requirements.txt
streamlit
opencv-python-headless
mediapipe
tensorflow==2.12.0
numpy
pandas
scikit-learn
googletrans==4.0.0-rc1
gTTS
pydub
Pillow
